requirements.txt
streamlit

###For RAG
langchain
python-dotenv
ipykernel
langchain-community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
duckdb
pandas
openai
langchain-groq
duckduckgo-search
pymupdf
arxiv
wikipedia
mysql-connector-python
SQLAlchemy
validators==0.28.1
youtube_transcript_api
unstructured
pytube
numexpr
huggingface_hub
langchain-ollama
duckduckgo-search
google-search-results
langdetect

streamlit-authenticator
google-auth
google-auth-oauthlib
google-auth-httplib2
python-dotenv
# sqlite3
firebase-admin
pyrebase4
transformers